server/pyproject.toml
[project]
name = "tradingagent"
version = "0.1.0"
description = "A generative AI agent for stock trading analysis and mock transactions."
authors = [
    { name = "Sudhanshu", email = "sudhanshu.sinha@think41.com" },
]
readme = "README.md"
requires-python = ">=3.10"
# Core dependencies for the application
dependencies = [
    "google-adk[mcp]==1.13.0",
    "fastapi-mcp",
    "yfinance==0.2.56",
    "psutil==5.9.5",
    "litellm==1.66.3",
    "google-generativeai==0.8.5",
    "python-dotenv==1.1.0",
    "Deprecated>=1.2.14",
    "vercel-ai",
    "fastapi[standard]",
    "uvicorn[standard]",
    "groq",
    "sqlalchemy",
    "psycopg2-binary",
    "alembic",
    "passlib",
    "bcrypt",
    "python-jose[cryptography]",
    "pydantic-settings",
    "pandas",
    "pandas_ta",
    "setuptools",
    "numpy<2.0",
    "mcp>=1.13.1",
    "httpx==0.23.3",
]

[project.optional-dependencies]
# Development dependencies for linting, formatting, and testing
# Install with: uv pip install -e ".[dev]"
dev = [
    "pytest>=8.0",
    "black>=24.0",
    "ruff>=0.4",
]

[build-system]
# Standard build system configuration
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent"]
